The first "almost" complete set of ARMv7 mnemonics and their variants
Like indicated in the README, the aim of this project is to provide
various files, that are easy to parse and provide various informations
about the elements composing the ARM assembly language, and the
instructions supported by this whole architecture set.

The current focus in on ARMv7 and ARMv8, but I'll gladly accept files
that provide informations concerning ARMv6 and lower architectures.

Signed-off-by: Myy <myy@miouyouyou.fr>